requirements.txt
django
#uwsgi
djangorestframework
markdown
django-filter
django-captcha-admin
vobject
requests

#For DJRadicale
Radicale